backend_py/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
# authors = ["Feodor Lebedev <lebedew.feodor@gmail.com>"]
readme = "README.md"

requires-python = ">= 3.12, < 4.0"
dependencies = [
    "fastapi ~= 0.115.0",
    "uvicorn ~= 0.31.0",
    "pydantic-settings ~= 2.9.0",
    "loguru ~= 0.7.2",
    "psycopg2-binary ~= 2.9.10",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose[cryptography] ~= 3.3.0",
    "passlib[bcrypt] ~= 1.7.4",
    "python-multipart ~= 0.0.6",
    "email-validator ~= 2.0.0",
    "httpx ~= 0.27.0",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "minio (>=7.2.16,<8.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "apscheduler (>=3.10.4,<4.0.0)",
]

[dependency-groups]
dev = [
    "ruff",
    "black"
]

upload = [
    "cyclopts",
    "httpx"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[tool.poetry.group.dev.dependencies]
ruff = "*"
black = "*"

[tool.poetry.group.upload.dependencies]
cyclopts = "*"
httpx = "*"

[tool.ruff]
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["F", "E", "W", "C90", "I", "S", "BLE", "T20", "ARG", "ERA", "PLR0913"]
ignore = ["E501", "S104", "C901", "BLE001"]

[tool.ruff.lint.pylint]
max-args = 7

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]
"scripts/extract_module_templates.py" = ["E402"]
"scripts/extract_schemas.py" = ["E402"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
